mocking for now

+++ parfait-api/src/main/java/com/megahack/parfait_api/service/ProductService.java
@@ -13,7 +13,7 @@ public interface ProductService extends BaseService<Product, ProductDto> {
 									  String terms, 
 									  String brand, 
 									  String category, 
-									  long lowestPrice, 
-									  long highestPrice);
+									  float lowestPrice, 
+									  float highestPrice);
 	List<String> getAvailableCategories();
 }
